refactor: header moved to Main directory
